+++ src/variables/particle.hpp
@@ -14,36 +14,16 @@ public:
 	bool spiderDashed = false;
 
 	class Drag {
-	private:
-		CCDictionary* dict =
-			CCDictionary::createWithContentsOfFile("dragEffect.plist");
-
-		CCPoint sourceStartSize = CCPoint(
-			((CCString*)dict->objectForKey("startParticleSize"))->floatValue(),
-			((CCString*)dict->objectForKey("startParticleSizeVariance"))->floatValue()
-		);
-
-		CCPoint sourceSpeed = CCPoint(
-			((CCString*)dict->objectForKey("speed"))->floatValue(),
-			((CCString*)dict->objectForKey("speedVariance"))->floatValue()
-		);
 	public:
 		static std::array<Drag, 2>& get() {
 			static std::array<Drag, 2> instance;
 			return instance;
 		}
-
-		float speed = sourceSpeed.x;
-		float speedVar = sourceSpeed.y;
-		float startSize = sourceStartSize.x;
-		float startSizeVar = sourceStartSize.y;
-
-		CCPoint sourcePosVar = CCPoint(
-			((CCString*)dict->objectForKey("sourcePositionVariancex"))->floatValue(),
-			((CCString*)dict->objectForKey("sourcePositionVariancey"))->floatValue()
-		);
-
-		CCPoint posVar = sourcePosVar;
+		float speed = 0.f;
+		float speedVar = 0.f;
+		float startSize = 0.f;
+		float startSizeVar = 0.f;
+		CCPoint posVar = CCPoint(0, 0);
 	};
 
 	class Trail {
@@ -53,9 +33,9 @@ public:
 			return instance;
 		}
 
-		CCPoint posVar = CCPoint(0, 2);
-		float speed = 15.f;
-		float speedVar = 4.f;
+		CCPoint posVar = CCPoint(0, 0);
+		float speed = 0.f;
+		float speedVar = 0.f;
 	};
 
 	class ShipClick {
@@ -65,11 +45,11 @@ public:
 			return instance;
 		}
 
-		CCPoint posVar = CCPoint(0, 2);
-		float speed = 150.f;
-		float speedVar = 40.f;
-		float startSize = 6.f;
-		float startSizeVar = 4.5f;
+		CCPoint posVar = CCPoint(0, 0);
+		float speed = 0.f;
+		float speedVar = 0.f;
+		float startSize = 0.f;
+		float startSizeVar = 0.f;
 	};
 
 	class Extra {
@@ -89,5 +69,9 @@ extern std::array<Particle::Drag, 2>& drag;
 extern std::array<Particle::Trail, 2>& trail;
 extern std::array<Particle::ShipClick, 2>& shipClick;
 extern std::array<Particle::Extra, 2>& extra;
-
-extern bool isInLevel;
+extern CCPoint sourceAngle;
+extern CCPoint sourceGravity;
+extern CCPoint sourceStartSize;
+extern CCPoint sourceSpeed;
+extern CCPoint sourcePosVar;
+extern std::array<int, 2> angleTweak;