+++ src/variables/settings.hpp
@@ -3,7 +3,7 @@
 extern bool toggle;
 extern bool legacyTracking;
 extern bool legacyScaling;
-extern bool inconstVal;
+extern bool isTrueVals;
 extern bool noRotation;
 extern bool centerDash;
 extern bool centerDashX;